refactor: extract relative time getting logic
